Kronborg

+++ goEmbeddedApi/api/models/startup.go
@@ -1,13 +1,14 @@
 package models
 
+// here we have a the Starup model that the orm use to create a startup table
 type Startup struct {
-	Id            int    `json:"id" gorm:"primaryKey"`
-	Local         string `json:"local" gorm:"unique"`
-	StartTime     int    `json:"start_time"`
-	EndTime       int    `json:"end_time"`
-	MinTemp       int    `json:"min_temp"`
-	MaxTemp       int    `json:"max_temp"`
-	MinHum        int    `json:"min_hum"`
-	MaxHum        int    `json:"max_hum"`
-	MaxNoiseLevel int    `json:"max_noise_level"`
+	Id            int     `json:"id" gorm:"primaryKey"`
+	Local         string  `json:"local" gorm:"unique"`
+	StartTime     int     `json:"start_time"`
+	EndTime       int     `json:"end_time"`
+	MinTemp       float64 `json:"min_temp"`
+	MaxTemp       float64 `json:"max_temp"`
+	MinHum        float64 `json:"min_hum"`
+	MaxHum        float64 `json:"max_hum"`
+	MaxNoiseLevel float64 `json:"max_noise_level"`
 }
